+++ packages/firebase-auth/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @fiboup/firebase-auth
 
+## 1.0.3
+
+### Patch Changes
+
+- [firebase-auth]
+
+  - export DecodedIdToken type
+
+  [hono-firebase-auth]
+
+  - allow developer to custom the current user object payload
+  - use an empty string for the current user context key to disable setting Hono current user context variable
+
 ## 1.0.2
 
 ### Patch Changes